Reorganize policies related command under a single point
Create aliases for shortcuts on some of them (applied and update), but
put them all under the single "policy" command category.
